Add visits and dns stacks
This change parameterizes the existing stack environments by
a single `domain` variable. This way, you can easily build
an array of stacks outside of a pipeline (for instance in dev
accounts) or build the stack directly into a pipeline, only
relying on the stage and environment as parameters.

Further, this change adds a stack called `MakerspaceDns` that
contains the Route53 public hosted zones for CloudFront/TLS
integration; plus the stack `Visitors` that is responsible
for the lambda and frontend console to register visits to the
makerspace.

Finally, the database has been split out into its own stack
so it can be shared across all of the different components of
the makerspace system. This also reduces the blast radius of
accidentally deleting an application stack, and it encourages
use of single-table-design, which is the best approach for
scaling usage in DynamoDB and for decreasing compute costs
related to spikes in traffic.

A non-code change to add a diagram of the future components
of the system in graphvis (and the PNG generated by it) is
included for convenience.