Build API Gateway for visits backend
This change includes some minor formatting changes.

The bulk of the important part of the commit is adding a new
class SharedApiGateway that routes all of the API traffic to
the appropriate Lambda functions. This is going to replace
the maintenance API gateway as well (eventually).

The purpose of the gateway is to sit behind a new DNS record
`api.cumaker.space`, which could eventually even be exposed to
students who could build their own integrations with the
makerspace.