Add makerspace stack to pipeline, add DNS to prod
The makerspace stack is going to replace the maintenance stack for now, and
we'll eventually add the maintenance app back in as a dependency of the
makerspace stage. This change removes the maintenance app from the top level
and from the pipeline, and adds in the maintenance app in its place.

It also turns the makerspace app into a stage for the Pipeline with the
help of a wrapper class, while leaving the original stack class in place
for the purpose of dev stages. I've filled out all the dev stages for the
existing capstone team.

I also added a DNS record for the visitors site. I'll fix more routes and
add the API route later, but for now I just want to get the basics out
so we can show a demo to the makerspace.