Update cloudfront and api gateway domains
Cloudfront is really easy to add a domain for. You just point the cloudfront
distribution with the ARecord Route53 construct. API Gateway is a little
trickier because you have to create a certificate to point at the time of
creation of the API Gateway itself, which references the domain name used
by the hosted zone.