Split Route53 zones and records
This is necessary in order to use Route53 for both CloudFront and API Gateway,
that's because both services will handle the certificates differently.
CloudFront creates its own certificate but will add your domain as an
alternative name. API Gateway, on the other hand, requires you to create the
certificate beforehand.

In order to create the certificate with API Gateway, we've got to prove
ownership of the domain, which means we must have created the hosted zone already
to manually enter the verification code.

CloudFront is simpler, but we can still take advantage of this setup to use
the authoritative domain name in setting the CloudFront domain, but also use
the CloudFront alternative names to set the Route53 record.

Without this change, there would be cross-stage and circular dependencies.
These changes split up the DNS concerns so that each stack owns its own.