Conditionally create DNS records to avoid cert validation
This is not what I wanted this change to be: it would be a lot cooler if
synthesizing in a development account created the records just like prod
or beta, but that would require creating an NS record in the production DNS
for every new dev account deploy, which stinks.

So instead I just selectively create DNS records here.