[BUGFIX] Deploy the Pipeline stack to the correct account
We had previously decided to deploy the Pipeline to the Beta account, and
while it was done that way before, I (without thinking) changed it to Prod.
That caused the pipeline to deploy, but it can't actually run the self-mutate
step because I had this changed back locally and not committed to GitHub
where it can make the changes.

This reverts that mistake.