Retrieve domain information from attributes
Rather than passing the built value of the cloudfront distribution
and Route53 domains directly to the Route53 ARecord objects, this
change pulls in the references and re-constructs the values from
the referenced objects' attributes. The difference is that in the
background, CDK will pull in a special `Mapping` value,
`AWSCloudFrontPartitionHostedZoneIdMap`, which is required in order
to re-construct the zone id. That mapping exists in the stack that
creates the CloudFront distribution, but not in the stack that
references.

After this change, the required mapping exists in the correct
stack.

I also took the liberty to fix the dependency graph so that all
the stacks underneath MakerspaceStack will be dependencies of it.
Without this, things still worked, but logically this is the correct
organization, and it could prevent issues in the future