Add autopep8 automatic GitHub actions rule
This rule acts only after the Flake8 rule to ensure that the
lint is already successful. Once that's true, we run
autopep8 in-place to enforce PEP-8 compliant syntax, that
way users don't have to run it themselves.

This doesn't ensure that the formatting of the code is
*good* per-se; that is still something we should require
of submitters, but this way we don't have to also ask
them to run the formatter.