test(cdk): add curl tests to CodePipeline for Beta and Prod urls
Problem:

In order to make sure any new code changes didn't break our front-end,
we can curl the Cloudfront urls to ensure we get a response.

Solution:

Added a Shellstep to both the Beta and Prod stages in Pipeline.py that
run a simple curl command on the Cloudfront endpoints hosting the login
system.

Testing:

This testing shellstep was deployed in a dev CodePipeline and ran
successfully when the dev endpoint was running.

Issue:
N/A, No open issue.