test(cdk/visit): adds API testing lambda and python script for later integration with pipeline and canary testing.
Problem:

In order to increase the amount of automated testing, we can verify our
API Gateway, Lambda, and Cloudfront services are functional with API
test scripts. We need to be able to automatically run these tests
whenever a new change is pushed and run tests continuously as a canary.

Solution:

This change adds a "TestAPILambda" function to the visit stack and
creates a testing_script.py. These files run the same tests, but the
lambda to be used in a canary test (Eventbridge trigger that can run the
lambda on some set timeframe) while the python script will be called in
a CodePipeline step. The tests will curl the proper frontend, call the
visit/register endpoints with dummy data (CANARY_TEST and PIPELINE_TEST
prefixes for the usernames), and verify that each endpoint is
functional. Note that this commit only adds the code to the
repo/Cloudformation. Later commits/changes will set up the tests to
actually run.

Testing:

The lambda and script were added to Dev visit stacks and CodePipelines
and displayed the correct behavior for situations where it should both
succeed and fail.

Issue:
N/A, No open issue.